requirements.txt
jupyter
numpy>=1.19.0
tensorflow>=2.2.0
pandas>=1.0.5
pillow
imagehash
matplotlib>=3.2.2
